Each process in linux has limit for file descriptors, the count of file descriptors created in a process can`t be biggest than the number in limits.
When we execute bfshell tool, then switchd creates a thread and file
descriptors for working with the bfshell. When the count of file
descriptors reaches a limit then we break from a loop. As result, we
can`t connect to switchd from bfshell in the future.

So now we are adding a handle for the case when the count of file
descriptors reaches the limit.